Update link pages - link community page - link user page
