--- src/main/java/com/cracker/controller/TestController.java
@@ -1,15 +1,0 @@
-package com.cracker.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class TestController {
-
-    @GetMapping("/")
-    public String hello(Model model){
-//        model.addAttribute("data", "hello!!");
-        return "home";
-    }
-}
