--- src/main/resources/static/js/placeSearch.js
@@ -1,8 +1,0 @@
-Kakao.init('ad824900200bc3d432c9b27e3ac74cd4');
-
-console.log(Kakao.isInitialized());
-
-function get_address() {
-    let matjip_n
-
-}
