+++ src/main/resources/templates/fragments/homeHeader.html
@@ -14,6 +14,7 @@
 
 
     <!-- Import JS -->
-    <script src="/js/map-place.js"></script>
+    <script src="/js/map.js"></script>
+    <script src="/js/place.js"></script>
 
 </head>
