+++ src/main/resources/templates/fragments/navbar.html
@@ -3,7 +3,7 @@
 <nav class="navbar is-fixed-top is-white" role="navigation" aria-label="main navigation" th:fragment="navbar">
     <div class="navbar-brand">
         <a class="navbar-item" href="/">
-            <img src="logo/logo_sub.png">&nbsp&nbsp
+            <img src="../logo/logo_sub.png">&nbsp&nbsp
             <h5 class="logo">CRACKER</h5>
             <strong class="is-sparta"
                     style="font-family: 'Stylish', sans-serif;font-size: larger;"></strong>
