+++ src/main/resources/templates/fragments/userHeader.html
@@ -10,6 +10,7 @@
     <link href="/css/user.css" rel="stylesheet">
 
     <!-- Import JS -->
-    <script src="/js/map-place.js"></script>
+    <script src="/js/place.js"></script>
+    <script src="/js/user.js"></script>
 
 </head>
